+++ package.json
@@ -2,10 +2,8 @@
   "name": "react-graphql-gh-vanilla",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://tmstani23.github.io/react-graphql-gh-vanilla",
   "dependencies": {
     "axios": "^0.18.0",
-    "gh-pages": "^2.0.1",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-scripts": "2.1.1"
@@ -14,8 +12,6 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
