+++ src/App.jsx
@@ -2,73 +2,27 @@
 
 import React, { Component } from 'react';
 import axios from 'axios';
-import Organization from './Components.jsx';
 
 const TITLE = 'React GraphQL GitHub Client';
-
-const ADD_STAR = `
-  mutation($repositoryId: ID!) {
-    addStar(input:{starrableId: $repositoryId}) {
-      starrable {
-        viewerHasStarred
-      }
-    }
-  }
-`;
-const REMOVE_STAR = `
-  mutation($repositoryId: ID!) {
-    removeStar(input:{starrableId: $repositoryId}) {
-      starrable {
-        viewerHasStarred
-      }
-    }
-  }
-`;
-
-const ADD_REACTION = `
-mutation($issueId: ID!){
-  addReaction(input:{subjectId:$issueId, content:HEART}) {
-    reaction {
-      content
-    }
-  }
-}
-`
+//Graphql query to the github api
 const GET_ISSUES_OF_REPOSITORY = `
-  query ($organization: String!, $repository: String!, $cursor: String) {
+  query ($organization: String!, $repository: String!) {
     organization(login: $organization) {
       name
       url
       repository(name: $repository) {
-        id
         name
         url
-        stargazers {
-          totalCount
-        }
-        viewerHasStarred
         licenseInfo{
           name
           description
         }
-        issues(first: 5, after: $cursor, states: [OPEN]) {
+        issues(last: 5) {
           edges {
             node {
               id
               title
               url
-              reactionGroups{
-                viewerHasReacted
-              }
-              reactions(last: 5) {
-                
-                edges {
-                  node {
-                    id
-                    content
-                  }
-                }
-              }
               comments(last: 5) {
                 edges {
                   node {
@@ -79,17 +33,93 @@ const GET_ISSUES_OF_REPOSITORY = `
               }
             }
           }
-          totalCount
-          pageInfo {
-            endCursor
-            hasNextPage
-          }
         }
       }
     }
   }
 `;
-
+//Component that displays the errors if any or returns
+  //the organization name and url.  A repository component is also displayed
+  //note: the organization/errors arguments are objects and are destructured giving access to properties when passing into the component
+const Organization = ({organization, errors}) => {
+  if (errors) {
+    return (
+      <p>
+        <strong>Something went wrong:</strong>
+        {errors.map(error => error.message).join(' ')}
+      </p>
+    )
+  }
+  return (
+  <div>
+    <p>
+      <strong>Issues from Organization: </strong>
+      <a href={organization.url}> {organization.name} </a>
+    </p>
+    <Repository repository={organization.repository} />
+  </div>
+  );
+}
+//Displays the repository name and a link to the repository
+const Repository = ({repository}) => {
+  //Display license info if available in the repository data object
+  if (repository.licenseInfo) {
+    return (
+      <div>
+        
+          <strong> Licenses:</strong> 
+          <p>{repository.licenseInfo.name}</p>
+          <p>{repository.licenseInfo.description}</p>
+        
+        <Issues repository = {repository} />
+      </div>
+    )
+  }
+  if (repository)
+  return (
+    <Issues repository = {repository} />
+  )
+}
+const Issues = ({repository}) => {
+  return (
+    // A link to the repository is displayed
+    <div>
+      <p>
+        <strong> In Repository:</strong>  
+        <a href={repository.url} > {repository.name} </a>
+      </p>
+    {/* each of the first five issues in the repo are mapped as list elements */}
+      <ul>
+        {repository.issues.edges.map((issue) => (
+          <li key={issue.node.id}>  
+            {/* links containing the issue title and url are displayed */}
+            <a href={issue.node.url}>{issue.node.title}</a>
+            <HasComment comment={issue.node.comments} />
+            <ul>
+              {/* For each issue map through the first 5 comments */}
+              {issue.node.comments.edges.map((comment) => (
+                // Display each comment body as a list element
+                <li key={comment.node.id}>
+                  <p>{comment.node.body}</p>
+                </li>
+              ))}
+            </ul>
+            {/* {console.log(issue.node.comments)} */}
+          </li>
+      ))}
+      </ul> 
+    </div>
+  )
+}
+//This component checks if there are comments available
+const HasComment = ({comment}) => {
+  //If there are comments display a comment message
+  return !comment.edges.length == 0 
+    ? ( <div>
+        <strong>Comments:</strong>  
+      </div> )
+    : null; 
+};
 //An axios http request object is created which includes the github api endpoint
         //and the requester's github authorization token
 const axiosGitHubGraphQL = axios.create({
@@ -101,142 +131,21 @@ const axiosGitHubGraphQL = axios.create({
     }`,
   },
 });
-//Component to query gh api using path and cursor location
-//Cursor is a hashed string representing a location within the repository issues
-const getIssuesofRepository = (path, cursor) => {
+
+const getIssuesofRepository = path => {
   //The input path is split at the '/' and each half is saved into separate arrays
   const [organization, repository] = path.split('/');
   //The axios post request to github is passed the query and gql variables
   return axiosGitHubGraphQL.post('', {
     query: GET_ISSUES_OF_REPOSITORY,
-    variables: { organization, repository, cursor },
-    })
+    variables: { organization, repository },
+    });
 };
-//This function returns an object containing the organization and errors data as well as handling issue merging after pagination.
-const resolveIssuesQuery = (queryResult, cursor) => state => {
-  const {data, errors} = queryResult.data;
-  //If there is no cursor(more issue button hasn't been clicked yet) return the organization data.
-  if(!cursor) {
-    return {
-      organization: data.organization,
-      errors,
-    }
-  }
-  //Create three variables containing the current state issue data, new issue data(that coming after the cursor), and an array containing both.
-  const {edges: oldIssues} = state.organization.repository.issues;
-  const {edges: newIssues} = data.organization.repository.issues;
-  const updatedIssues = [...oldIssues, ...newIssues];
-  //Return a nested object containing the current organization, repo and issue data.
-    //The issues edges are where to parse from all issue data and is passed the updated issue range
-    //This object is used to update the respective state within the App component.
-  return {
-    organization: {
-      ...data.organization,
-      repository: {
-        ...data.organization.repository,
-        issues: {
-          ...data.organization.repository.issues,
-          edges: updatedIssues,
-        },
-      },
-    },
-    errors,
-  };
-};
-//function that handles making the add star mutation post to the GH api
-const addStarToRepository = repositoryId => {
-  return axiosGitHubGraphQL.post('', {
-    query: ADD_STAR,
-    variables: { repositoryId },
-  });
-};
-
-//function that handles making the add star mutation post to the GH api
-const removeStarToRepository = repositoryId => {
-  return axiosGitHubGraphQL.post('', {
-    query: REMOVE_STAR,
-    variables: { repositoryId },
-  });
-};
-
-const addReactionToIssue = issueId => {
-  
-  console.log(issueId);
-  return axiosGitHubGraphQL.post('', {
-    query: ADD_REACTION,
-    variables: { issueId },
-  })
-}
-
-// const resolveAddReactionMutation = mutationResult => state => {
-//   //Save current viewHasStarred status to a variable
-//   const {
-//     content,
-//   } = mutationResult.data.data.addReaction.reaction;
-  
-  
-//   return {
-//     ...state,
-//       organization: {
-//         ...state.organization,
-//           repository: {
-//             ...state.organization.repository,
-//             issues {
-//               ...state.organization.issues,
-//             }
-//           }
-      
-//       }
-  
-// }
-
-const resolveAddStarMutation = mutationResult => state => {
- //Save current viewHasStarred status to a variable
-  const {
-    viewerHasStarred,
-  } = mutationResult.data.data.addStar.starrable;
-  //Create totalCount variable containing current # of stargazers
-  const { totalCount } = state.organization.repository.stargazers
-  
-  return {
-    ...state,
-    organization: {
-      ...state.organization,
-      repository: {
-        ...state.organization.repository,
-        viewerHasStarred,
-        stargazers: {
-          // total count is incremented +1 after starring and updated within the nested object structure
-          totalCount: totalCount + 1,
-        }
-      },
-    },
-  }
-}
-const resolveRemoveStarMutation = mutationResult => state => {
-  //Save current viewHasStarred status to a variable
-   const {
-     viewerHasStarred,
-   } = mutationResult.data.data.removeStar.starrable;
-   //Create totalCount variable containing current # of stargazers
-   const { totalCount } = state.organization.repository.stargazers
-   
-   return {
-     ...state,
-     organization: {
-       ...state.organization,
-       repository: {
-         ...state.organization.repository,
-         viewerHasStarred,
-         stargazers: {
-           // total count is incremented +1 after starring and updated within the nested object structure
-           totalCount: totalCount - 1,
-         }
-       },
-     },
-   }
- }
-
+//This function returns an object containing the organization and errors data
+const resolveIssuesQuery = queryResult => () => ({
+  organization: queryResult.data.data.organization,
+  errors: queryResult.data.errors,
+});
 //Main App component which handles rendering, lifecycle and user input.
 class App extends Component {
   state = {
@@ -260,44 +169,12 @@ class App extends Component {
     //prevent default submit button behavior
     event.preventDefault();
   };
-  //Make request to github api and update the App state with the returned data
-  onFetchFromGitHub = (path, cursor)=> {
-    //Call query method and update organization state with the returned nested data object(organization,repository,issues).
-    getIssuesofRepository(path, cursor).then(queryResult => 
-      this.setState(resolveIssuesQuery(queryResult, cursor)),  
+  onFetchFromGitHub = path => {
+
+    getIssuesofRepository(path).then(queryResult => 
+      this.setState(resolveIssuesQuery(queryResult)),  
     );
   };
-  //This method takes the current endCursor location from the state and calls the onFetchFromGithub chain.
-  onFetchMoreIssues = () => {
-    //Cursor location is destructured from the organization state object and passed to the fetch method.
-    const {
-      endCursor,
-    } = this.state.organization.repository.issues.pageInfo;
-    
-    this.onFetchFromGitHub(this.state.path, endCursor)
-  }
-
-  //Method that calls the add star query post to the GH api
-  onStarRepository = (repositoryId, viewerHasStarred) => {
-    if(viewerHasStarred) {
-      removeStarToRepository(repositoryId).then(mutationResult => {
-        this.setState(resolveRemoveStarMutation(mutationResult))
-      });
-    }
-    else {
-      addStarToRepository(repositoryId).then(mutationResult => {
-        this.setState(resolveAddStarMutation(mutationResult))
-      });
-    }
-    
-    
-  }
-  onReactionToIssue = issueId => {
-    addReactionToIssue(issueId).then(result => {
-      console.log(result);
-    })
-  }
-
   render() {
     const { path, organization, errors } = this.state;
 
@@ -309,7 +186,6 @@ class App extends Component {
           <label htmlFor="url">
             Show open issues for https://github.com
           </label>
-          {/* Input field that contains the organization and repository */}
           <input
             id="url"
             type="text"
@@ -321,15 +197,8 @@ class App extends Component {
         </form>
 
         <hr />
-        {/* Once data is returned display the Organization component with the returned data. 
-          onFetchmoreIssues (cursor location and page info) is passed to the component for use down the chain*/}
         {organization ? (
-          <Organization 
-            organization={organization} 
-            errors={errors} 
-            onFetchMoreIssues={this.onFetchMoreIssues} 
-            onStarRepository={this.onStarRepository}
-            onReactionToIssue={this.onReactionToIssue}/>
+          <Organization organization={organization} errors = {errors} />
         ) : (
           <p>No Information Yet...</p>
         )}
